+++ test/runtests.jl
@@ -66,6 +66,7 @@ using GLLVM
     include("test_beta_fit.jl")
     include("test_gamma_fit.jl")
     include("test_tweedie.jl")
+    include("test_tweedie_engine_health.jl")
     include("test_exponential.jl")
     include("test_studentt.jl")
     include("test_lognormal.jl")